+++ src/lib/bars/index.ts
@@ -11,6 +11,7 @@ export * from './chart-bar-grouped';
 export * from './chart-bar-stacked';
 export * from './series-label';
 export * from './series-label-bar';
+export * from './chart-window-bar';
 export * from './chart-window-bar-grouped';
 
 // export * from './grouped-bars';